+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Released: 2020-01-26
  * Handling sensor chip subfeature not being readable
  * HWMON: displaying average value as current if current is not available
  * Added clarification text on how to read RAM info
+ * Added check to verify if `dmidecode` is available on the host
 
 Version 0.6.1
 =============